+++ src/openneuro/_config.py
@@ -7,7 +7,7 @@ from typing import TypedDict
 
 import platformdirs
 
-from openneuro._logging import log
+from openneuro._logging import _unicode, log
 
 CONFIG_DIR = Path(
     platformdirs.user_config_dir(appname="openneuro-py", appauthor=False, roaming=True)
@@ -27,8 +27,12 @@ class Config(TypedDict):
 def init_config() -> None:
     """Initialize a new OpenNeuro configuration file."""
     log(
-        "🙏 Please login to your OpenNeuro account and go to: "
-        "My Account → Obtain an API Key"
+        _unicode(
+            "Please login to your OpenNeuro account and go to: "
+            "My Account → Obtain an API Key",
+            emoji="🙏",
+            end="",
+        )
     )
     api_key = getpass.getpass("OpenNeuro API key (input hidden): ")
